trusted-types/TrustedTypes-createPolicy.tentative.html

<!DOCTYPE html>
<script src="/resources/testharness.js" ></script>
<script src="/resources/testharnessreport.js"></script>
<script src="support/helper.sub.js"></script>

<meta http-equiv="Content-Security-Policy" content="trusted-types *">
<body>
<script>
  // Adopted from @koto's polyfill tests
  test(t => {
    const p = TrustedTypes.createPolicy('policy', {});

    assert_true(p.createHTML instanceof Function);
    assert_true(p.createScript instanceof Function);
    assert_true(p.createScriptURL instanceof Function);
    assert_true(p.createURL instanceof Function);
    assert_equals(p.name, 'policy');
  }, 'TrustedTypes.createPolicy returns a working policy object.');

  test(t => {
    let p = TrustedTypes.createPolicy('frozencheck', {});

    // Assigning a new property
    p.a = 'foo';
    assert_false('foo' in p);

    // Overwriting an existing property
    let oldCreateHTML = p.createHTML;
    p.createHTML = (s) => s;
    assert_equals(oldCreateHTML, p.createHTML);

    // Deleting a property
    delete p.createHTML;
    assert_true('createHTML' in p);
  }, 'TrustedTypes.createPolicy returns an unmodifiable policy object.');

  test(t => {
    const p = TrustedTypes.createPolicy('empty', {});

    assert_throws(new TypeError(), _ => {
      p.createHTML('foo');
    });

    assert_throws(new TypeError(), _ => {
      p.createScript('foo');
    });

    assert_throws(new TypeError(), _ => {
      p.createScriptURL('foo');
    });

    assert_throws(new TypeError(), _ => {
      p.createURL('foo');
    });
  }, 'TrustedTypes.createPolicy() with empty policy - createXXX reject by default.');

  test(t => {
    assert_throws(new TypeError(), _ => {
      const p = TrustedTypes.createPolicy('policy', null);
    });
  }, 'TrustedTypes.createPolicy() with null throws.');
</script>
